# test(consensus): Print the `consensus_time_to_receive_block` metric at the end of `consensus_performance_test`