I think it's a useful metrics to keep track of.
Also, added a small test case where we send very few, very small messages.

```
2024-09-18 14:55:22.612 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG: ===== Summary for //rs/tests/consensus:consensus_performance_colocate =====
2024-09-18 14:55:22.612 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG: Task setup                   PASSED  in 126.30s   -- Exited with code 0.
2024-09-18 14:55:22.612 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG: Task test_few_small_messages PASSED  in  98.22s
2024-09-18 14:55:22.612 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Success rate: 100.0%
2024-09-18 14:55:22.616 INFO[setup:rs/tests/testing_verification/colocate_test.rs:321:0] Test execution finished with exit code 0.
2024-09-18 14:55:22.616 INFO[setup:rs/tests/testing_verification/colocate_test.rs:239:0] Wait extra 10 sec to collect last uvm logs
2024-09-18 14:55:22.725 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Block rate: 1.2 blocks/s
2024-09-18 14:55:22.725 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Throughput: 0.0 MiB/s, 1.0 messages/s
2024-09-18 14:55:22.725 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Average time to receive a rank 0 block: 0.2s
2024-09-18 14:55:22.725 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Avarage E2E ingress message latency: 2.2s
2024-09-18 14:55:22.725 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG: Task test_small_messages     PASSED  in 104.33s
2024-09-18 14:55:22.725 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Success rate: 100.0%
2024-09-18 14:55:22.725 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Block rate: 0.5 blocks/s
2024-09-18 14:55:22.725 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Throughput: 1.5 MiB/s, 354.1 messages/s
2024-09-18 14:55:22.725 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Average time to receive a rank 0 block: 1.9s
2024-09-18 14:55:22.725 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Avarage E2E ingress message latency: 14.1s
2024-09-18 14:55:22.725 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG: Task test_large_messages     PASSED  in 115.55s
2024-09-18 14:55:22.725 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Success rate: 100.0%
2024-09-18 14:55:22.725 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Block rate: 0.5 blocks/s
2024-09-18 14:55:22.725 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Throughput: 1.2 MiB/s, 1.3 messages/s
2024-09-18 14:55:22.725 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Average time to receive a rank 0 block: 2.0s
2024-09-18 14:55:22.725 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Avarage E2E ingress message latency: 15.3s
2024-09-18 14:55:22.725 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG: ===========================================================================
```